Step2.
A user page has been created with the ability to edit. Redax and React hooks are used to control the state. Redus-thunk is used for dispatching asynch actions.
Grid and Material-UI are used for layout and design. In the third step, the ability to search for users and pagination will be added.